rename generators.py to buffers.py
